requirements.txt
aiogram>=3.0
aiogram-dialog>=2.1
apscheduler>=3.10
boto3>=1.34
pandas>=2.2
pyarrow>=15,<16
requests>=2.31
python-dotenv>=1.0
trino[auth]>=0.328
urllib3>=2.2
dagster>=1.5